everything is back to working order... though this doesn't change the functionality, the code looks MUCH better and is easier to follow. For the most part at least. Things still got a bit cloudy towards user sign out.
