+++ client/src/App.js
@@ -1,10 +1,8 @@
 // Import React
 import React, { Component } from 'react';
 // Import Components
-// import NavDialogs from './components/Dialogs';
 import Navbar from './components/Navbar';
 
-// import SignUp from './components/dialogs/SignUp';
 // Import CSS
 import './App.css';
 
@@ -21,11 +19,8 @@ class App extends Component {
   render() {
     return (
       <div className="App">
-        {/* I Pass in the Sign In/Sign Up dialog, Which in turn passes in the Navbar. */}
-        {/* This way, the Navbar can open the dialogs and have access to them in the future*/}
-        {/* <NavDialogs /> */}
+        {/* I Pass in the Sign In/Sign Up dialogs thru the Navbar */}
         <Navbar />
-        {/* <SignUp /> */}
 
         <h1>We puts the stuffs here!</h1>
       </div>
